+++ packages/react/src/prefabs/PreJoin.tsx
@@ -11,6 +11,7 @@ import {
   facingModeFromLocalTrack,
   Track,
   VideoPresets,
+  Mutex,
 } from 'livekit-client';
 import * as React from 'react';
 import { MediaDeviceMenu } from './MediaDeviceMenu';
@@ -64,28 +65,40 @@ export function usePreviewTracks(
 ) {
   const [tracks, setTracks] = React.useState<LocalTrack[]>();
 
+  const trackLock = React.useMemo(() => new Mutex(), []);
+
   React.useEffect(() => {
-    let trackPromise: Promise<LocalTrack[]> | undefined = undefined;
     let needsCleanup = false;
-    if (options.audio || options.video) {
-      trackPromise = createLocalTracks(options);
-      trackPromise
-        .then((tracks) => {
+    trackLock.lock().then(async (unlock: () => void) => {
+      try {
+        tracks?.forEach((track) => {
+          track.stop();
+        });
+        if (options.audio || options.video) {
+          const localTracks = await createLocalTracks(options);
+
           if (needsCleanup) {
-            tracks.forEach((tr) => tr.stop());
+            localTracks.forEach((tr) => tr.stop());
           } else {
-            setTracks(tracks);
+            setTracks(localTracks);
           }
-        }).catch(onError);
-    }
+        }
+      } catch (e: unknown) {
+        if (onError && e instanceof Error) {
+          onError(e);
+        } else {
+          log.error(e);
+        }
+      } finally {
+        unlock();
+      }
+    });
 
     return () => {
       needsCleanup = true;
-      trackPromise?.then((tracks) =>
-        tracks.forEach((track) => {
-          track.stop();
-        }),
-      );
+      tracks?.forEach((track) => {
+        track.stop();
+      });
     };
   }, [JSON.stringify(options)]);